requirements.txt
asgiref==3.9.2
Django==5.2.7
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
PyJWT==2.10.1
sqlparse==0.5.3
tzdata==2025.2
mysqlclient==2.2.1        # if using MySQL
psycopg2-binary==2.9.9   # if using PostgreSQL
python-dotenv==1.0.0      # optional, for env variables
